requirements.txt
Flask>=2.0.0,<3.0.0
python-dotenv>=0.19.0,<1.0.0
Flask-WTF>=1.0.0,<2.0.0
Flask-Login>=0.5.0,<0.7.0 # Added Flask-Login
email-validator>=1.1.0,<2.0.0
stripe>=3.0.0,<4.0.0
requests>=2.25.0,<3.0.0
Flask-Moment>=1.0.0,<2.0.0 # Added Flask-Moment
